modified:   python_introduction/finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 # calculate finance related values
-monthly_income = int(input("Enter your monthly income: "))
-total_mExpenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+total_mExpenses = float(input("Enter your total monthly expenses: "))
 monthly_savings = monthly_income - total_mExpenses 
 
 # annual savings
